Reset player experience and level when cancelling lobby coutdown
When a player exit a game lobby, and the lobby start coutdown needs to
be cancelled, remaining player's experience and level is no longer
updated and seems to freeze.
This commit resets players experience and level to a maximum value
(100% experience and level 999) when the lobby timer is inactive and/or
cancelled.